requirements.txt
tensorflow==2.18.0
gradio>=4.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.2
Pillow>=10.0.0
flask>=2.2.0